Move feedback view tests to their own file
